requirements.txt
Flask
flask-bcrypt
flask-cors
google-generativeai
PyMuPDF
psycopg2-binary
pgvector
sqlalchemy
python-dotenv
requests
Pillow
markdown
